BUG: Ensure compile errors are raised correclty
This has been bugging me for a bit.  The concurrent.futures Executor
requires checking the result for the error to be raised.  That makes
sense, but just means we need to consume the result explicitly here
to ensure we know about compile errors.

Otherwise, compile errors just pass silently (which is very confusing
if the old object files are still around and the tests run based on
the old version).