PERF: Fast check on equivalent arrays in PyArray_EQUIVALENTLY_ITERABLE_OVERLAP_OK (#21464)
Addresses one of the items in #21455

* fast check on equivalent arrays in PyArray_EQUIVALENTLY_ITERABLE_OVERLAP_OK

* fix for case stride1==0

* remove assert statements

* whitespace

* DOC: Add comment that output broadcasting is rejected by overlap detection

That is a bit of a weird dynamic here, and it would probably nice to clean
up.  However, it is also not a reason to not add the fast-path right now.

---

Note that it may also be good to just use a data-pointer comparison or push this fast-path into the overlap detection function itself.