DOC: improve the docstring of numpy.sinc to explain behavior at zero.
The first lines of the sinc docstring did not mention the behavior
when the argument is zero, and gave the impression that the function is
defined as the quotient between a sine and its argument everywhere. That
information was instead relegated to the "Notes" section. I found this
issue while navigating the documentation for jax.numpy.sinc, which
inherits those few lines. I think it is important to document not only
the value of the function at zero, but also that it make sinc a smooth
function. This connects better with some of the other routes used to
introduce sinc (e.g. as a Taylor series). I have tried to implement
the conclusions of the discussion around Issue #21538, keeping the
description close to the actual implementation and avoiding my original
suggestion of introducing the function as a limit.