API: Retain `arr.base` more strictly in `np.frombuffer`
Checking the implementation of `releasebuffer` is the most strict
version for when we can safely assume that wrapping into a
`memoryview` is not necessary.
In theory, `view.obj` could also be set to a new object even with
the old buffer-protocol, in practice I don't think that should
happen (the docs read like it should not be used).

So this is the minimal fix to retain old behavior as much as
(safely) possible.

Closes gh-21612